# server/openapi.yaml
...
openapi: 3.0.3
info:
  title: GWP API
  version: 0.1.0
  description: >
    GWP — Canvas Online as a self-hosted app. Phase 0 ships first-run
    setup, session auth and health probes; Phase 1 adds the organization
    surface (company profile, departments, teams — spec §3/§4). Canvas,
    coaching and AI endpoints arrive in later phases.
  contact:
    name: GWP operators
servers:
  - url: /
    description: Same origin as the served web app.
...
paths:
...
  /api/v1/coaching-sessions:
    get:
      tags: [coaching]
      summary: List the caller's coaching sessions
      description: >
        Lets a session without a report be resumed after a reload. Scope
        mirrors the existing session relation and never widens it: an
        owner sees every session of the company; anyone else only sessions
        where they are the coach of record or the creator. The coachee,
        other managers, admin and readers of a linked canvas see nothing
        (spec §6). Metadata only. Ordered newest-created first over
        (created_at, id); limit defaults to 25, max 100; cursor is the
        opaque token from nextCursor.
      operationId: listCoachingSessions
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - name: cursor
          in: query
          required: false
          description: Opaque keyset cursor from the previous page's nextCursor.
          schema:
            type: string
      responses:
        "200":
          description: One page of visible sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSessionPage"
        "400":
          description: Invalid limit or cursor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [coaching]
      summary: Record a coaching session
      description: >
        The declared coach is the LOGGED-IN actor and must currently manage
        the coachee (manager role + current reporting subtree) — or be the
        owner, who may coach anyone. An owner may also name a different
        coachUserId to enter a historical session on that coach's behalf;
        the override is audited as `coaching.session.create_on_behalf` and
        the creator flag never grants share rights (spec §6). Naming
        another coach as a non-owner is 403. `coacheeUserId` must be an
        ACTIVE company user different from the coach; `occurredAt` may not
        be in the future. An optional `canvasId` links the session to a
        canvas the actor can read under the subject policy — checked
        independently, since report rights never flow from canvas rights.
        Sessions are records of fact: no update or delete route exists.
      operationId: createCoachingSession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachingSessionCreateRequest"
      responses:
        "201":
          description: Session recorded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingSessionCreateResponse"
        "400":
          description: >
            Malformed body — coach equals coachee, occurredAt in the
            future, or an invalid field type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Non-owner naming a different coachUserId.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: >
            The coach is not the coachee's current manager, the coachee
            or named coach does not exist in this company, or the canvas
            is missing/unreadable — the same denial for every case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: The coachee is not an active user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/reports:
    get:
      tags: [coaching]
      summary: List coaching reports the caller may read
      description: >
        Scoped by the report ACL (spec §6) — report creator, session
        coach, owner, or holder of an active report_share — and never by
        the reporting tree or canvas access. `coacheeUserId` filters
        inside the visible scope. Ordered newest-first over
        (created_at, id); limit defaults to 25, max 100; cursor is the
        opaque token from nextCursor. List items carry metadata only —
        `body` and `provenance` are served by GET /reports/{id}.
      operationId: listCoachingReports
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - name: cursor
          in: query
          required: false
          description: Opaque keyset cursor from the previous page's nextCursor.
          schema:
            type: string
        - name: coacheeUserId
          in: query
          required: false
          description: Restrict the page to sessions for one coachee.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: One page of visible report list items.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReportPage"
        "400":
          description: Invalid limit, cursor, or coacheeUserId.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      tags: [coaching]
      summary: Save a validated ORACLE preview as an immutable report
      description: >
        Explicit save (spec §6) — the ONLY way a grader preview becomes a
        report. The request carries ids and an idempotency key, never a
        report body: the server re-reads its own validated in-memory
        preview, re-checks CURRENT write authority (owner, or the session
        coach who still manages the coachee — checked again at save time,
        not just run start), and inserts the next immutable
        `report_version` under the company lock. Provenance is copied onto
        the report row so retention may later delete the `ai_run`
        (FK ON DELETE SET NULL) without losing it. Replaying the same
        idempotencyKey returns the same report; a re-grade under a new key
        appends a new version. Audited as `coaching.report.save` with
        metadata only — body and transcript never enter audit or logs.
      operationId: saveCoachingReport
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoachingReportSaveRequest"
      responses:
        "201":
          description: Report version created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReportSaveResponse"
        "200":
          description: Idempotent replay — same key, same ids.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReportSaveResponse"
        "400":
          description: Invalid body shape (unknown/malformed fields).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Related actor who lost current report-write
            authority (e.g. the coach no longer manages the coachee).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Session or run missing, foreign, or owned by
            another user — uniform, no existence disclosure.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: AI_RUN_MISMATCH (not this session's oracle run) or
            AI_RUN_NOT_SUCCEEDED.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "410":
          description: PREVIEW_EXPIRED — the validated preview is gone
            (TTL or restart); grade again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: ORACLE_REPORT_INVALID — the preview is not a
            validated grader output.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/reports/{id}:
    get:
      tags: [coaching]
      summary: Read one coaching report
      description: >
        The same ACL as the list: creator, session coach, owner, or an
        active share. Missing, foreign-company, and unreadable ids all
        return the same 404. Every authorized read is audited as
        `coaching.report.read` with metadata only — the report body never
        enters audit or logs.
      operationId: getCoachingReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Report id inside the caller's company.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The report.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingReport"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Missing, foreign, or not shared with the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [coaching]
      summary: Delete a report (confirmed, body + shares)
      description: >
        Coach of record or owner only — decided AFTER the read ACL, so
        unreadable ids still answer 404 while a readable non-manager gets
        403. The body MUST be `{"confirm": true}` — anything else is 400.
        Removes the report row and cascades its shares; the audit trail
        keeps `coaching.report.delete` with metadata only — deleted
        content never survives in audit.
      operationId: deleteCoachingReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirm]
              additionalProperties: false
              properties:
                confirm:
                  type: boolean
                  enum: [true]
      responses:
        "204":
          description: Deleted — body and shares gone.
        "400":
          description: Missing or false confirm flag.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Readable but not coach-of-record or owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Missing, foreign, or not shared with the caller.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/reports/{id}/shares/{userId}:
    put:
      tags: [coaching]
      summary: Share one report version with a company user
      description: >
        Explicit share grant (spec §6) on THIS report row — a new
        report_version gets a fresh empty share set. Coach of record or
        owner only (read ACL first: unreadable → 404, readable
        non-manager → 403). The target must be an active same-company
        user; sharing a user who already holds ACL access is harmless.
        Re-sharing a revoked grant re-activates it.
      operationId: shareCoachingReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          description: Same-company active user to grant read access.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Share active (fresh or re-activated).
          content:
            application/json:
              schema:
                type: object
                required: [reportId, userId]
                properties:
                  reportId:
                    type: string
                    format: uuid
                  userId:
                    type: string
                    format: uuid
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Readable but not coach-of-record or owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Report unreadable/missing, or target user unknown
            or inactive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [coaching]
      summary: Revoke a report share
      description: >
        Same manage gate as the grant. Revocation is effective from the
        next request; revoking a grant that does not exist (or is already
        revoked) is a plain 404.
      operationId: revokeCoachingReportShare
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Share revoked.
        "401":
          description: Missing/expired/revoked session or inactive user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Readable but not coach-of-record or owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Report unreadable/missing, or no active share.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  parameters:
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Must equal the `gwp_csrf` cookie (double-submit).
      schema:
        type: string
    LimitParam:
      name: limit
      in: query
      required: false
      description: Page size — default 25, maximum 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Stable machine code (e.g. INVALID_SESSION).
        message:
          type: string
          description: Human-readable, Vietnamese UI copy.
        details:
          description: Optional structured detail (field names, pending list).
        request_id:
          type: string
          format: uuid
          description: Echo of the X-Request-Id response header.
...
    CoachingSessionCreateRequest:
      type: object
      required: [coachUserId, coacheeUserId, occurredAt]
      additionalProperties: false
      properties:
        coachUserId:
          type: string
          format: uuid
          description: >
            The logged-in actor — a different value is a historical entry
            on that coach's behalf, which only an owner may record.
        coacheeUserId:
          type: string
          format: uuid
          description: Active company user inside the coach's current subtree.
        canvasId:
          type: string
          format: uuid
          description: Optional canvas link — must be readable by the actor.
        occurredAt:
          type: string
          format: date-time
          description: When the session happened — not in the future.
    CoachingSessionCreateResponse:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
    CoachingSessionListItem:
      type: object
      required:
        [id, coachUserId, coacheeUserId, canvasId, occurredAt, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        coachUserId:
          type: string
          format: uuid
        coacheeUserId:
          type: string
          format: uuid
        canvasId:
          type: string
          format: uuid
          nullable: true
        occurredAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    CoachingSessionPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CoachingSessionListItem"
        nextCursor:
          type: string
          nullable: true
          description: Pass as ?cursor= for the next page; null at the end.
    CoachingReportListItem:
      type: object
      required:
        - id
        - sessionId
        - coachUserId
        - coacheeUserId
        - canvasId
        - reportVersion
        - rubricVersion
        - createdBy
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        sessionId:
          type: string
          format: uuid
        coachUserId:
          type: string
          format: uuid
        coacheeUserId:
          type: string
          format: uuid
        canvasId:
          type: string
          format: uuid
          nullable: true
        reportVersion:
          type: integer
        rubricVersion:
          type: string
          nullable: true
          description: e.g. ORACLE-v3 — null until a grader run fills it.
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    CoachingReportPage:
      type: object
      required: [items, nextCursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CoachingReportListItem"
        nextCursor:
          type: string
          nullable: true
          description: Pass as ?cursor= for the next page; null at the end.
    CoachingReport:
      allOf:
        - $ref: "#/components/schemas/CoachingReportListItem"
        - type: object
          required: [occurredAt, aiRunId, body, provenance]
          properties:
            occurredAt:
              type: string
              format: date-time
            aiRunId:
              type: string
              format: uuid
              nullable: true
              description: The grader run that produced the body, if any.
            body:
              type: object
              description: >
                The validated grader report document — never the raw
                transcript, which is never stored.
            provenance:
              type: object
              description: Grader provenance metadata (model, rubric, timing).
    CoachingReportSaveRequest:
      type: object
      required: [sessionId, runId, idempotencyKey]
      additionalProperties: false
      description: >
        Explicit save — ids only, never a report body. The body is read
        from the server's own validated preview; a client-supplied report
        is not accepted.
      properties:
        sessionId:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
          description: The caller's own succeeded oracle run on sessionId.
        idempotencyKey:
          type: string
          minLength: 8
          maxLength: 200
          description: Same key + same ids replay the original report.
    CoachingReportSaveResponse:
      type: object
      required: [reportId, reportVersion]
      additionalProperties: false
      properties:
        reportId:
          type: string
          format: uuid
        reportVersion:
          type: integer
          description: Immutable per-session version, assigned under the
            company lock.